src/ChatBot/ChatBot/agent/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.4.0
bedrock-agentcore>=0.1.0
boto3>=1.35.0
